Refactored: replaced addQPTRef() with addQPTRefs() and determineBGPPosition().
* addQPTRefs() now takes into consideration "abbreviated" sparql, i.e.,
the use of ';' and ','. The latter is still under test, though. As a
result, one Entity IRI can be part of more BGP's, hence produces more
QPT Refs. 
* determineBGPPosition(): generic code, extracted from original
addQPTRef, to determine for a ParseStruct node its {s,p,o} position in
the BGP triple. Returns None on failure to determine (e.g., when node is
not part of query pattern part of query)