Refactored addQPTRefs w/t getTriples: extruded the downwards flow to getTriples().
* addQPTRefs was very complicated since it mixed upwards and downwards
directions when (i) searching for triple parts, and (ii) completing the
parts with upwards element(s) property and/or subject.
* addQPTRefs now only addresses the upward flow, i.e.,
object->property->subject, completing triples. This part is
straightforward, no recursion required anymore.
* New method getTriples(): only addresses downwards flow, i.e., from
current node, find all triples that share this node. This is a recursive
algorithm, but rather straightforward.